// src/types/NFTContract.ts
export default interface NFTContractType {
   collection: {
      banner_image_url: string
      chat_url: string
      created_date: string
      description: string
      discord_url: string
      external_url: string
      safelist_request_status: string
      image_url: string
      is_subject_to_whitelist: boolean
      large_image_url: string
      medium_username: string
      name: string
      slug: string
      telegram_url: string
      twitter_username: string
      instagram_username: string
      wiki_url: string
      is_nsfw: boolean
   }
   address: string
   total_supply: string
}